requirements.txt
# TruthSim Requirements
# ====================

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# HTTP Requests (for UMLS API)
requests>=2.31.0

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0
together>=1.0.0

# For local model inference (optional)
transformers>=4.38.0
torch>=2.0.0
accelerate>=0.27.0
bitsandbytes>=0.42.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Evaluation Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Logging
loguru>=0.7.0

# Jupyter (for analysis)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
